requirements.txt
absl-py==0.8.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
future==0.18.2
google-auth==1.6.3
google-auth-oauthlib==0.4.1
grpcio==1.24.3
gym==0.15.3
idna==2.8
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.3
oauthlib==3.1.0
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycolab==1.2
pyglet==1.3.2
pyparsing==2.4.2
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scipy==1.3.1
six==1.12.0
tensorboard==2.0.1
torch==1.3.0.post2
torchvision==0.4.1.post2
urllib3==1.25.6
Werkzeug==0.16.0
